pyproject.toml
[project]
name = "restro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.16.5",
    "fastapi>=0.116.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
    "cryptography"
]
